pinkycode-feature-added_comment_classifier/requirements.txt
openai>=1.88.0
groq>=0.29.0
textstat>=0.7.0
nltk>=3.9.1
psycopg2-binary>=2.9.10
requests>=2.32.4
pylint>=2.17.7
pytest>=7.0
pytest-cov>=4.0.0
torch
transformers